+++ src/main/java/mx/com/nmp/pagos/mimonte/dto/TipoContactoRespDTO.java
@@ -1,32 +1,34 @@
 package mx.com.nmp.pagos.mimonte.dto;
+
 /**
- * Nombre: ContactoRespDTO Descripcion: Clase que encapsula la informacion perteneciente a los tipo de contactos.
+ * Nombre: TipoContactoRespDTO Descripcion: Clase que encapsula la informacion
+ * perteneciente a los tipo de contactos.
  *
  * @author José Rodríguez jgrodriguez@quarksoft.net
  * @creationDate 06/03/2019 16:27 hrs.
  * @version 0.1
  */
 public class TipoContactoRespDTO {
-	
-	private Integer id;
-	
+
+	private Long id;
+
 	private String descripcion;
 
 	public TipoContactoRespDTO() {
 		super();
 	}
 
-	public TipoContactoRespDTO(Integer id, String descripcion) {
+	public TipoContactoRespDTO(Long id, String descripcion) {
 		super();
 		this.id = id;
 		this.descripcion = descripcion;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
